+++ CommonTest/TicksTest.cpp
@@ -19,14 +19,14 @@ public:
 		Assert::AreEqual(m.timestamp, sut.timestamp);
 		Assert::AreEqual(m.price, sut.prices().front());
 		Assert::AreEqual(m.price, sut.prices().data()[0]);
-		Assert::AreEqual(m.volume, (size_t)sut.volumes().front());
-		Assert::AreEqual(m.volume, (size_t)sut.volumes().data()[0]);
+		Assert::AreEqual(m.volume, sut.volumes().front());
+		Assert::AreEqual(m.volume, sut.volumes().data()[0]);
 
 		// when timestamp is same, subsequent request should update price and add volume
 		m.price *= 2;
 		Assert::IsFalse(sut.AddTick(m));
 		Assert::AreEqual(m.price, sut.prices().back());
-		Assert::AreEqual(m.volume * 2, (size_t)sut.volumes().back());
+		Assert::AreEqual(m.volume * 2, sut.volumes().back());
 	}
 
 	TEST_METHOD(PutTest_Multiple) {
